--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/shiron-dev/rapi/cmd"
-
-func main() {
-	cmd.Execute()
-}
